+++ src/main/java/com/orders/OrderObjcetRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import com.orders.OrderObjcet;
 //public interface OrderObjcetRepository extends CrudRepository<OrderObjcet, Long> {
 
-public interface OrderObjcetRepository extends CrudRepository<OrderObjcet, Integer> {
+public interface OrderObjcetRepository extends CrudRepository<OrderObjcet, Long> {
 }
